LC-2052-Implement user locking logic
