+++ src/main/java/uk/gov/cabinetoffice/csl/service/IdentityService.java
@@ -1,0 +1,112 @@
+package uk.gov.cabinetoffice.csl.service;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.annotation.ReadOnlyProperty;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import uk.gov.cabinetoffice.csl.domain.*;
+import uk.gov.cabinetoffice.csl.exception.IdentityNotFoundException;
+import uk.gov.cabinetoffice.csl.exception.ResourceNotFoundException;
+import uk.gov.cabinetoffice.csl.exception.UnableToAllocateAgencyTokenException;
+import uk.gov.cabinetoffice.csl.repository.IdentityRepository;
+import uk.gov.cabinetoffice.csl.service.client.csrs.ICivilServantRegistryClient;
+
+import java.time.Instant;
+import java.util.HashSet;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
+
+@Slf4j
+@Service
+@Transactional
+public class IdentityService {
+
+    private final InviteService inviteService;
+    private final AgencyTokenCapacityService agencyTokenCapacityService;
+    private final IdentityRepository identityRepository;
+    private final ICivilServantRegistryClient civilServantRegistryClient;
+    private final PasswordEncoder passwordEncoder;
+
+    public IdentityService(InviteService inviteService,
+                           AgencyTokenCapacityService agencyTokenCapacityService,
+                           IdentityRepository identityRepository,
+                           ICivilServantRegistryClient civilServantRegistryClient,
+                           PasswordEncoder passwordEncoder) {
+        this.inviteService = inviteService;
+        this.agencyTokenCapacityService = agencyTokenCapacityService;
+        this.identityRepository = identityRepository;
+        this.civilServantRegistryClient = civilServantRegistryClient;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Transactional(noRollbackFor = {UnableToAllocateAgencyTokenException.class, ResourceNotFoundException.class})
+    public void createIdentityFromInviteCode(String code, String password, TokenRequest tokenRequest) {
+        Invite invite = inviteService.getInviteForCode(code);
+        String email = invite.getForEmail();
+        final String domain = getDomainFromEmailAddress(email);
+        Set<Role> newRoles = new HashSet<>(invite.getForRoles());
+        String agencyTokenUid = null;
+        if (tokenRequest != null && tokenRequest.hasData()) {
+            Optional<AgencyToken> agencyTokenForDomainTokenOrganisation =
+                    civilServantRegistryClient.getAgencyTokenForDomainTokenOrganisation(tokenRequest.getDomain(),
+                            tokenRequest.getToken(), tokenRequest.getOrg());
+
+            agencyTokenUid = agencyTokenForDomainTokenOrganisation
+                    .map(agencyToken -> {
+                        if (agencyTokenCapacityService.hasSpaceAvailable(agencyToken)) {
+                            return agencyToken.getUid();
+                        } else {
+                            throw new UnableToAllocateAgencyTokenException("Agency token uid " +
+                                    agencyToken.getUid() + " has no spaces available. Identity not created");
+                        }
+                    })
+                    .orElseThrow(ResourceNotFoundException::new);
+
+            log.info("Identity request has agency uid = {}", agencyTokenUid);
+        } else if (!isAllowListedDomain(domain) && !inviteService.isEmailInvited(email)) {
+            log.info("Invited request neither agency, nor allowListed, nor invited via IDM: {}", invite);
+            throw new ResourceNotFoundException();
+        }
+
+        Identity identity = new Identity(
+                UUID.randomUUID().toString(),
+                email,
+                passwordEncoder.encode(password),
+                true,
+                false,
+                newRoles,
+                Instant.now(),
+                false,
+                agencyTokenUid,
+                0);
+
+        identityRepository.save(identity);
+
+        log.debug("New identity email = {} successfully created", email);
+    }
+
+    @ReadOnlyProperty
+    public boolean isIdentityExistsForEmail(String email) {
+        return identityRepository.existsByEmailIgnoreCase(email);
+    }
+
+    public Identity getIdentityForEmail(String email) {
+        return identityRepository.findFirstByEmailEqualsIgnoreCase(email);
+    }
+
+    public Identity getIdentityForEmailAndActiveFalse(String email) {
+        return identityRepository
+                .findFirstByActiveFalseAndEmailEqualsIgnoreCase(email)
+                .orElseThrow(() -> new IdentityNotFoundException("Identity not found for email: " + email));
+    }
+
+    public String getDomainFromEmailAddress(String emailAddress) {
+        return emailAddress.substring(emailAddress.indexOf('@') + 1);
+    }
+
+    public boolean isAllowListedDomain(String domain) {
+        return civilServantRegistryClient.getAllowListDomains().contains(domain.toLowerCase());
+    }
+}
