+++ src/main/java/uk/gov/cabinetoffice/csl/validation/validator/IsCurrentPasswordValidator.java
@@ -5,7 +5,7 @@ import jakarta.validation.ConstraintValidatorContext;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.cabinetoffice.csl.domain.Identity;
-import uk.gov.cabinetoffice.csl.service.UserService;
+import uk.gov.cabinetoffice.csl.service.PasswordService;
 import uk.gov.cabinetoffice.csl.service.auth2.IUserAuthService;
 import uk.gov.cabinetoffice.csl.validation.annotation.IsCurrentPassword;
 
@@ -13,13 +13,13 @@ import uk.gov.cabinetoffice.csl.validation.annotation.IsCurrentPassword;
 @AllArgsConstructor
 public class IsCurrentPasswordValidator implements ConstraintValidator<IsCurrentPassword, String> {
 
-   private final UserService userService;
+   private final PasswordService passwordService;
    private final IUserAuthService userAuthService;
 
    public void initialize(IsCurrentPassword constraint) {}
 
    public boolean isValid(String value, ConstraintValidatorContext context) {
       Identity identity = userAuthService.getIdentity();
-      return userService.checkPassword(identity.getEmail(), value);
+      return passwordService.checkPassword(identity.getEmail(), value);
    }
 }
