+++ src/test/java/uk/gov/cabinetoffice/csl/controller/reset/ResetControllerTest.java
@@ -12,8 +12,8 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.cabinetoffice.csl.domain.Identity;
 import uk.gov.cabinetoffice.csl.domain.Reset;
+import uk.gov.cabinetoffice.csl.service.IdentityService;
 import uk.gov.cabinetoffice.csl.service.ResetService;
-import uk.gov.cabinetoffice.csl.service.UserService;
 import uk.gov.cabinetoffice.csl.util.TestUtil;
 
 import java.util.Date;
@@ -48,7 +48,7 @@ public class ResetControllerTest {
     private ResetService resetService;
 
     @MockBean
-    private UserService userService;
+    private IdentityService identityService;
 
     @Test
     public void shouldLoadRequestResetForm() throws Exception {
@@ -67,7 +67,7 @@ public class ResetControllerTest {
 
     @Test
     public void shouldLoadCheckEmailPageIfIdentityExistsForTheGivenEmailId() throws Exception {
-        when(userService.isIdentityExistsForEmail(EMAIL)).thenReturn(true);
+        when(identityService.isIdentityExistsForEmail(EMAIL)).thenReturn(true);
         mockMvc.perform(post("/reset")
                         .param("email", EMAIL)
                         .with(csrf())
@@ -86,7 +86,7 @@ public class ResetControllerTest {
 
     @Test
     public void shouldLoadRequestResetFormWithErrorMessageIfIdentityDoesNotExistForTheGivenEmailId() throws Exception {
-        when(userService.isIdentityExistsForEmail(EMAIL)).thenReturn(false);
+        when(identityService.isIdentityExistsForEmail(EMAIL)).thenReturn(false);
         mockMvc.perform(post("/reset")
                         .param("email", EMAIL)
                         .with(csrf())
@@ -200,7 +200,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "Password")
                         .param("confirmPassword", "Password")
@@ -232,7 +232,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "PASSWORD123")
                         .param("confirmPassword", "PASSWORD123")
@@ -264,7 +264,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "password123")
                         .param("confirmPassword", "password123")
@@ -296,7 +296,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "Pass123")
                         .param("confirmPassword", "Pass123")
@@ -328,7 +328,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "Password123")
                         .param("confirmPassword", "PasswordMisMatch")
@@ -357,7 +357,7 @@ public class ResetControllerTest {
         when(resetService.getResetByCode(CODE)).thenReturn(reset);
         when(resetService.isResetExpired(reset)).thenReturn(false);
         when(resetService.isResetComplete(reset)).thenReturn(false);
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(null);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(null);
 
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "Password123")
@@ -384,7 +384,7 @@ public class ResetControllerTest {
 
         Identity identity = TestUtil.createIdentity(ID, UID, EMAIL, PASSWORD, null);
 
-        when(userService.getIdentityForEmail(EMAIL)).thenReturn(identity);
+        when(identityService.getIdentityForEmail(EMAIL)).thenReturn(identity);
         mockMvc.perform(post("/reset/" + CODE)
                         .param("password", "Password123")
                         .param("confirmPassword", "Password123")
